matcher: move Matcher into it's own module
And:
* add support for loading patterns from JSON files
